Replace Template exported
Function to replace the template exported to the modules folder and imported into the main index file
