+++ index.js
@@ -1,6 +1,8 @@
 const fs = require('fs')
 const http = require('http')
 const url = require('url')
+
+const replaceTemplate = require('./modules/replaceTemplate')
 /////////////////////////////////////////////////
 //Reading the file previously to not read it everytime the endpoint gets hit and creating the templates
 const tempOverview = fs.readFileSync(`${__dirname}/templates/template-overview.html`, 'utf-8')
@@ -9,21 +11,6 @@ const tempCard = fs.readFileSync(`${__dirname}/templates/template-card.html`, 'u
 const data = fs.readFileSync(`${__dirname}/dev-data/data.json`,'utf-8')
 const productData = JSON.parse(data)
 //SERVER OPERATIONS
-//Function to replace the placeholders witht he information provided by the API
-const replaceTemplate = (temp, product) =>{
-    let output = temp.replace(/{%PRODUCTNAME%}/g,product.productName);
-    output = output.replace(/{%IMAGE%}/g,product.image)
-    output = output.replace(/{%QUANTITY%}/g,product.quantity)
-    output = output.replace(/{%PRICE%}/g,product.price)
-    output = output.replace(/{%DESCRIPTION%}/g,product.description)
-    output = output.replace(/{%FROM%}/g,product.from)
-    output = output.replace(/{%NUTRIENTS%}/g,product.nutrients)
-    output = output.replace(/{%ID%}/g,product.id)
-
-    if(!product.organic) output = output.replace(/{%NOT_ORGANIC%}/g,'not-organic')
-    
-    return output
-} 
 //Creating the server
 const server = http.createServer((req, res) => {
     const {query,pathname} = url.parse(req.url,true)
